requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
streamlit
